pedidos stages

+++ src/Pages/OneOrder/index.jsx
@@ -1,6 +1,6 @@
 /* eslint-disable jsx-a11y/alt-text */
 import { useState, useEffect } from "react";
-import { useParams } from "react-router-dom";
+import { useParams,useNavigate } from "react-router-dom";
 import api from "../../Services/api";
 import { getToken } from "../../Services/auth";
 import { toast } from "react-toastify";
@@ -19,16 +19,17 @@ const OneOrder = () => {
   const [pedido, setPedido] = useState({});
   const [itens, setItens] = useState([]);
   const [loading, setLoading] = useState(true);
+  const navigate = useNavigate();
 
   const total = itens.reduce((total, item)=>{
     return total + item.value;
   }, 0);
 
-
   useEffect(() => {
     try {
       async function loadPedido () {
         const response = await api.get(`orderAll/${id}`);
+        // console.log(response.data)
         setPedido(response.data);
         setItens(response.data.itensObject);
       }
@@ -41,15 +42,77 @@ const OneOrder = () => {
   }, [id]);
 
   const handleDelete = async () => {
-    setLoading(true);
+    if( window.confirm(`Você tem certeza disso? Não será possivel restalrar o pedido da mesa ${pedido.table}`)){
+      setLoading(true);
+      const headers = { 
+        'authorization': `Bearer ${getToken()}`,
+      };
+      try {
+        await api.delete(`order/${id}`, headers)
+        .then(() => {toast.success(`Pedido Mesa ${pedido.table}, deletado com sucesso`); navigate("/orders");})
+        .catch((res) => toast.error(`Algo deu errado, tente entrar novamente`));
+      }catch(err){
+        console.log(err);
+      } finally {
+        setLoading(false);
+      }
+    }
+  }
 
+  const handleAceitar = async () => {
+    setLoading(true);
     const headers = { 
       'authorization': `Bearer ${getToken()}`,
     };
     try {
-      await api.delete(`order/${id}`, headers)
-      .then(() => toast.success('Deletado com sucesso'))
-      .catch((res) => toast.warn('Algo deu errado, tente logar novamente'));
+      await api.patch(`order/${id}`, {confirmation: true} , headers)
+      .then(() => {
+        toast.success(`O Pedido Mesa ${pedido.table}, foi aceito e está esperando a confirmação da cozinha`);
+        let order = pedido;
+        order.confirmation = true;
+        setPedido(order);
+      })
+      .catch((res) => toast.error(`Algo deu errado, tente entrar novamente`));
+    }catch(err){
+      console.log(err);
+    } finally {
+      setLoading(false);
+    }
+  }
+  const handleComecar = async () => {
+    setLoading(true);
+    const headers = { 
+      'authorization': `Bearer ${getToken()}`,
+    };
+    try {
+      await api.patch(`order/${id}`, {inProgess: true} , headers)
+      .then(() => {
+        toast.success(`O Pedido Mesa ${pedido.table}, Começou a ser produzido`);
+        let order = pedido;
+        order.inProgess = true;
+        setPedido(order);
+      })
+      .catch((res) => toast.error(`Algo deu errado, tente entrar novamente`));
+    }catch(err){
+      console.log(err);
+    } finally {
+      setLoading(false);
+    }
+  }
+  const handleEntregar = async () => {
+    setLoading(true);
+    const headers = { 
+      'authorization': `Bearer ${getToken()}`,
+    };
+    try {
+      await api.patch(`order/${id}`, {isFinish: true} , headers)
+      .then(() => {
+        toast.success(`O Pedido Mesa ${pedido.table}, Foi entregue`);
+        let order = pedido;
+        order.isFinish = true;
+        setPedido(order);
+      })
+      .catch((res) => toast.error(`Algo deu errado, tente entrar novamente`));
     }catch(err){
       console.log(err);
     } finally {
@@ -110,18 +173,18 @@ const OneOrder = () => {
               </div>
               <div className="status-order">
                 <img src={pedido.confirmation ? Confirm : Stay} />
-                <img src={pedido.isProgress ? Confirm : Stay} />
+                <img src={pedido.inProgess ? Confirm : Stay} />
                 <img src={pedido.isFinish ? Confirm : Stay} />
               </div>
             </div>
             <div className="status-button"> 
               {
                 !pedido.confirmation ? (
-                  <><span>Aceitar</span><button className="btn"><img src={Check} /></button></>
+                  <><span>Aceitar</span><button className="btn" onClick={handleAceitar}><img src={Check} /></button></>
                 ) : !pedido.inProgess ? (
-                  <><span>Começar</span><button className="btn"><img src={Check} /></button></>
+                  <><span>Começar</span><button className="btn" onClick={handleComecar}><img src={Check} /></button></>
                 ) : pedido.inProgess && !pedido.isFinish ? (
-                  <><span>Entregar</span><button className="btn"><img src={Check} /></button></>
+                  <><span>Entregar</span><button className="btn" onClick={handleEntregar}><img src={Check} /></button></>
                 ) : !pedido.isPaid ? (
                   <><span>Fechar</span><button className="btn"><img src={Check} /></button></>
                 ) : "Pedido já foi finalizado"
